Progress concurrent events management
