+++ src/main/java/it/s3sync/sync/SynchornizationThread.java
@@ -22,9 +22,7 @@ import it.s3sync.listener.WatchListeners.Operation;
 import it.s3sync.model.Item;
 import it.s3sync.service.SynchronizationService;
 import it.s3sync.service.UploadService;
-import it.s3sync.sync.EventData.AmqpEventData;
 import it.s3sync.sync.EventData.EventSource;
-import it.s3sync.sync.EventData.FileSystemEventData;
 import it.s3sync.sync.EventData.S3Action;
 import it.s3sync.utils.FileUtils;
 import it.s3sync.utils.SpringContext;
